Remove no longer used Vim plugins
tslime.vim is for sending a portion of text from a vim buffer to a
running tmux session, but I don't normally use tmux, so this is
unnecessary.

I no longer use Cucumber in any active projects, so that plugin can be
removed.

tern_for_vim provides Tern-based JavaScript editing support, but the
project is no longer actively supported.

vim-turbux is designed to speed up your TDD cycle by using a single key
mapping to run the most relevant test based on the current file in vim.
I no longer use tmux, so this is unnecessary as well.